Set up communication by Lora

+++ src/PacketGPS.cpp
@@ -1,5 +1,5 @@
 #include "PacketGPS.h"
 
 PacketGPS::PacketGPS(int hh, int mm, int ss,
-			float lngG, float latG, int numSatG) :
+			double lngG, double latG, int numSatG) :
 	h(hh), m(mm), s(ss), lng(lngG), lat(latG), numSat(numSatG) {}
