+++ src/PacketGPS.h
@@ -6,11 +6,11 @@ public:
 	int m;
 	int s;
 
-	float lng;
-	float lat;
+	double lng;
+	double lat;
 	int numSat;
 	
 	PacketGPS(int hh = 0, int mm = 0, int ss = 0,
-			float lngG = 0, float latG = 0, int numSatG = 0);
+			double lngG = 0, double latG = 0, int numSatG = 0);
 
 };
